migrated to sonar production data endpoint, miinor module refactoring
