+++ .gitignore
@@ -1,3 +1,3 @@
 .DS_Store
 .idea
- __pycache__
+ __pycache__/*
