+++ Dockerfile
@@ -21,7 +21,13 @@ WORKDIR /app
 
 ENV AWS_ACCESS_KEY_ID=<AWS_ACCESS_KEY_ID>
 ENV AWS_SECRET_ACCESS_KEY=<AWS_SECRET_ACCESS_KEY>
+ENV DEBUG_ENV True
+ENV HTTP_UI_ADDRESS http://managerui:80
+ENV S3_ENDPOINT http://minio
 
+ENV MONGO_URL=mongodb
+ENV MONGO_PORT=27017
+ENV STAT_DB_NAME=hydrostat
 
 EXPOSE 5000
 CMD python app.py
