+++ hydro_stat/monitor_stats.py
@@ -1,8 +1,4 @@
-import statistics
-
 import numpy as np
-from loguru import logger
-from scipy.stats import entropy as ent
 
 
 # import numpy as np
